+++ cli/src/base.ts
@@ -29,32 +29,8 @@ import {
   readCredentialsDictionary
 } from './credentials.js';
 import { reportBugURL } from './utils.js';
+import { ProjectConfig, partialProjectConfig } from './config.js';
 
-export const projectConfigSchema = z.object({
-  databaseURL: z.string(),
-  codegen: z.object({
-    output: z.string(),
-    moduleType: z.enum(['cjs', 'esm', 'deno']),
-    declarations: z.boolean(),
-    javascriptTarget: z.enum([
-      'es5',
-      'es6',
-      'es2015',
-      'es2016',
-      'es2017',
-      'es2018',
-      'es2019',
-      'es2020',
-      'es2021',
-      'esnext'
-    ]),
-    workersBuildId: z.string().optional()
-  })
-});
-
-const partialProjectConfig = projectConfigSchema.deepPartial();
-
-export type ProjectConfig = z.infer<typeof partialProjectConfig>;
 export type APIKeyLocation = 'shell' | 'dotenv' | 'profile' | 'new';
 
 const moduleName = 'xata';
