--- cli/src/feature-flags.ts
@@ -1,9 +1,0 @@
-// eslint-disable-next-line
-const flags = (process.env.XATA_CLI_FLAGS || '')
-  .split(',')
-  .map((s) => s.toLocaleLowerCase().trim())
-  .filter(Boolean);
-
-export const features = {
-  //flagName: flags.includes('flagName')
-};
