+++ cli/src/utils/feature-flags.ts
@@ -1,0 +1,9 @@
+import { ProjectConfig } from '../config';
+
+export function hasExperimentalFlag(config: ProjectConfig, flag: keyof NonNullable<ProjectConfig['experimental']>) {
+  return config.experimental?.[flag] === true;
+}
+
+export function hasRemoteFlag() {
+  // TODO: implement
+}
